fixed bug with NULL A. added python module with CVXOPT compatibility

+++ src/runecos.c
@@ -83,5 +83,6 @@ int main(void)
         
     }
 	
-	return exitflag;
+    /* explicitly truncate exit code */
+	return (int)exitflag;
 }
